System: Make autobackup work + bugfixes, UI: Add backup settings page

+++ J-Runner/Panels/NandTools.cs
@@ -17,7 +17,7 @@ namespace JRunner.Panels
         {
             return (int)numericIterations.Value;
         }
-        public void setNumericIterations(decimal value)
+        public void setNumericIterations(int value)
         {
             numericIterations.Value = value;
         }
